(rough) changed implementation of Result<T> to be 8 bytes
I haven't checked the code gen on it. Technically it's not much if at all slower because the only case where extra logic has to be done is when there is an error. That's it, I've looked at some code gen comparing it to a implementation of std::expected and a basic "if the function fails return -1" implementation. Now, the results were great, the only time where extra code was generated was for the failure case, other than that the code gen was exactly the same! We don't care about the failure case, it's permissible to be slower because the failure case isn't likely. I haven't done the bool implementation yet just because of laziness. I'll do it later.